+++ app/app.go
@@ -1039,6 +1039,7 @@ func New(
 	}
 
 	// Currently in an upgrade hold for this block.
+	var storeLoader baseapp.StoreLoader
 	if upgradeInfo.Name != "" && upgradeInfo.Height == app.LastBlockHeight()+1 {
 		if app.UpgradeKeeper.IsSkipHeight(upgradeInfo.Height) {
 			app.Logger().Info("Skipping upgrade based on height",
@@ -1053,13 +1054,14 @@ func New(
 				"lastHeight", app.LastBlockHeight(),
 			)
 			// See if we have a custom store loader to use for upgrades.
-			storeLoader := GetUpgradeStoreLoader(app, upgradeInfo)
-			if storeLoader != nil {
-				app.SetStoreLoader(storeLoader)
-			}
+			storeLoader = GetUpgradeStoreLoader(app, upgradeInfo)
 		}
 	}
 	// --
+
+	// Verify configuration settings
+	storeLoader = ValidateWrapper(app.Logger(), appOpts, storeLoader)
+	app.SetStoreLoader(storeLoader)
 
 	if loadLatest {
 		if err := app.LoadLatestVersion(); err != nil {